+++ src/ss_math.h
@@ -13,7 +13,7 @@ float normal_dist (int num, int place);
 
 /** Returns the logistic of the requested value
  * 1/(1 + exp(-value)) */
-float logist(double value);
+double logist(double value);
